+++ src/Polly.Shared/Retry/RetryEngine.cs
@@ -1,7 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Runtime.ExceptionServices;
 using System.Threading;
+
+#if NET40
+using ExceptionDispatchInfo = Polly.Utilities.ExceptionDispatchInfo;
+#endif
 
 namespace Polly.Retry
 {
@@ -37,13 +42,20 @@ namespace Polly.Retry
                 }
                 catch (Exception ex)
                 {
-                    if (!shouldRetryExceptionPredicates.Any(predicate => predicate(ex)))
+                    Exception handledException = shouldRetryExceptionPredicates
+                        .Select(predicate => predicate(ex))
+                        .FirstOrDefault(e => e != null);
+                    if (handledException == null)
                     {
                         throw;
                     }
 
-                    if (!policyState.CanRetry(new DelegateResult<TResult>(ex), cancellationToken))
+                    if (!policyState.CanRetry(new DelegateResult<TResult>(handledException), cancellationToken))
                     {
+                        if (handledException != ex)
+                        {
+                            ExceptionDispatchInfo.Capture(handledException).Throw();
+                        }
                         throw;
                     }
                 }