+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 5.6.0
 - Bug fix: set context keys for generic execute method with PolicyWrap
+- Add ability to handle inner exceptions natively: .HandleInner&lt;TEx&gt;()
 - Add GetPolicies extension method to IPolicyWrap
 - Allow WaitAndRetry policies to calculate wait based on the handled fault